renamed delay line convenience handles to name_firstcell and name_lastcell
since with the scheme I had before with input/output, I feel you'd expect
numberofcells delay from input to output, but you're actually getting a delay
of (numberofcells-1).

The point of this software as I see it is to be able to know the contents of
each memory location at each moment of the simulation, so I've opted to keep
the naming explicit in terms of the layout of the data structures.